- create `MarkdownDimens` to also offer ability to change some additional values - expand `MarkdownPadding` with more options - restructure some code
