+++ multiplatform-markdown-renderer-m2/src/commonMain/kotlin/com/mikepenz/markdown/m2/Markdown.kt
@@ -14,8 +14,9 @@ fun Markdown(
     content: String,
     colors: MarkdownColors = markdownColor(),
     typography: MarkdownTypography = markdownTypography(),
-    padding: MarkdownPadding = markdownPadding(),
     modifier: Modifier = Modifier.fillMaxSize(),
+    padding: MarkdownPadding = markdownPadding(),
+    dimens: MarkdownDimens = markdownDimens(),
     flavour: MarkdownFlavourDescriptor = GFMFlavourDescriptor(),
     imageTransformer: ImageTransformer = ImageTransformerImpl(),
     components: MarkdownComponents = markdownComponents(),
@@ -23,8 +24,9 @@ fun Markdown(
     content,
     colors,
     typography,
-    padding,
     modifier,
+    padding,
+    dimens,
     flavour,
     imageTransformer,
     components,
