+++ multiplatform-markdown-renderer-m2/src/commonMain/kotlin/com/mikepenz/markdown/m2/MarkdownColors.kt
@@ -13,10 +13,12 @@ fun markdownColor(
     linkText: Color = text,
     codeBackground: Color = MaterialTheme.colors.onBackground.copy(alpha = 0.1f),
     inlineCodeBackground: Color = codeBackground,
+    dividerColor: Color = MaterialTheme.colors.onSurface.copy(alpha = 0.12f),
 ): MarkdownColors = DefaultMarkdownColors(
     text = text,
     codeText = codeText,
     linkText = linkText,
     codeBackground = codeBackground,
     inlineCodeBackground = inlineCodeBackground,
+    dividerColor = dividerColor
 )
